Add D3DBSP_Lib Implemention to D3DBSP_Converter

+++ D3DBSP_Lib/LightGridColors.cpp
@@ -35,7 +35,7 @@ int ConvertLump_BOtoWAW_LightGridColors(Lump* boLump, Lump* wawLump)
 
 	for(DWORD i = 0; i < LightGridColorCount * 56 * 3; i++)
 	{
-		dst[i] = src[i];
+		dst[i] = (BYTE)src[i];
 	}
 
 	return 0;
